refactor(desktop): split settings and automation surfaces
