+++ apps/desktop/src/components/automations/controls/AutomationRunLocationSelector.tsx
@@ -1,25 +1,26 @@
-import type { ReactNode } from "react";
-import { ComputeTargetSwatch } from "@/components/compute/ComputeTargetSwatch";
 import { PillControlButton } from "@proliferate/ui/primitives/PillControlButton";
-import { PopoverMenuItem } from "@proliferate/ui/primitives/PopoverMenuItem";
 import {
   POPOVER_SURFACE_CLASS,
   PopoverButton,
 } from "@proliferate/ui/primitives/PopoverButton";
 import {
-  Check,
   CloudIcon,
   FolderOpen,
-  Plus,
-  Terminal,
 } from "@proliferate/ui/icons";
 import type {
   AutomationTargetGroup,
-  AutomationTargetRow,
   AutomationTargetSelection,
 } from "@/lib/domain/automations/target/selection";
-
-type AutomationRunOwnerScope = "personal" | "organization";
+import {
+  findDefaultAutomationTargetRow,
+  findSelectedAutomationTargetRow,
+  renderAutomationTargetRowIcon,
+  RunLocationMenuItem,
+  RunLocationRows,
+  RunLocationSectionHeader,
+  type AutomationRunLocationConfigureTarget,
+  type AutomationRunOwnerScope,
+} from "@/components/automations/controls/run-location/AutomationRunLocationMenu";
 
 interface AutomationOwnerOption {
   value: AutomationRunOwnerScope;
@@ -38,16 +39,10 @@ interface AutomationRunLocationSelectorProps {
   disabledReason: string | null;
   onSelectOwner: (ownerScope: AutomationRunOwnerScope) => void;
   onSelectTarget: (target: AutomationTargetSelection) => void;
-  onConfigureCloud: (target: {
-    gitOwner: string;
-    gitRepoName: string;
-    ownerScope: AutomationRunOwnerScope;
-  }) => void;
+  onConfigureCloud: (target: AutomationRunLocationConfigureTarget) => void;
 }
 
 const RUN_LOCATION_SURFACE_CLASS = `w-72 min-w-[175px] ${POPOVER_SURFACE_CLASS}`;
-const RUN_LOCATION_SECTION_CLASS =
-  "flex min-h-6 items-center truncate px-2 py-1 text-sm leading-4 text-muted-foreground";
 const RUN_LOCATION_DIVIDER_CLASS = "mx-1 my-1.5 h-px scale-y-50 bg-foreground/10";
 
 export function AutomationRunLocationSelector({
@@ -64,13 +59,13 @@ export function AutomationRunLocationSelector({
 }: AutomationRunLocationSelectorProps) {
   const personalOption = ownerOptions.find((option) => option.value === "personal");
   const teamOption = ownerOptions.find((option) => option.value === "organization");
-  const selectedPersonalRow = findSelectedTargetRow(personalGroups);
-  const selectedTeamRow = findSelectedTargetRow(teamGroups);
-  const personalDefaultRow = selectedPersonalRow ?? findDefaultTargetRow(personalGroups);
+  const selectedPersonalRow = findSelectedAutomationTargetRow(personalGroups);
+  const selectedTeamRow = findSelectedAutomationTargetRow(teamGroups);
+  const personalDefaultRow = selectedPersonalRow ?? findDefaultAutomationTargetRow(personalGroups);
   const teamDefaultRow =
     selectedTeamRow
-    ?? findDefaultTargetRow(teamGroups, "cloud")
-    ?? findDefaultTargetRow(teamGroups);
+    ?? findDefaultAutomationTargetRow(teamGroups, "cloud")
+    ?? findDefaultAutomationTargetRow(teamGroups);
   const selectedRow = ownerScope === "organization" ? selectedTeamRow : selectedPersonalRow;
   const teamDisabledReason =
     teamOption?.disabledReason
@@ -87,10 +82,10 @@ export function AutomationRunLocationSelector({
     : selectedRow?.repoLabel ?? null;
   const triggerIcon = ownerScope === "organization"
     ? selectedRow
-      ? targetRowIcon(selectedRow, "trigger")
+      ? renderAutomationTargetRowIcon(selectedRow, "trigger")
       : <CloudIcon className="size-3.5" />
     : selectedRow
-    ? targetRowIcon(selectedRow, "trigger")
+    ? renderAutomationTargetRowIcon(selectedRow, "trigger")
     : <FolderOpen className="size-3.5" />;
   const activeGroups = ownerScope === "organization" ? teamGroups : personalGroups;
   const activeOption = ownerScope === "organization" ? teamOption : personalOption;
@@ -200,235 +195,4 @@ export function AutomationRunLocationSelector({
       ) : null}
     </div>
   );
-}
-
-function RunLocationSectionHeader({ label }: { label: string }) {
-  return (
-    <div className={RUN_LOCATION_SECTION_CLASS}>
-      {label}
-    </div>
-  );
-}
-
-function RunLocationRows({
-  activeOwnerScope,
-  disabledReason,
-  emptyLabel,
-  groups,
-  isLoading,
-  ownerScope,
-  onSelectOwner,
-  onSelectTarget,
-  onConfigureCloud,
-}: {
-  activeOwnerScope: AutomationRunOwnerScope;
-  disabledReason: string | null;
-  emptyLabel: string;
-  groups: AutomationTargetGroup[];
-  isLoading: boolean;
-  ownerScope: AutomationRunOwnerScope;
-  onSelectOwner: (ownerScope: AutomationRunOwnerScope) => void;
-  onSelectTarget: (target: AutomationTargetSelection) => void;
-  onConfigureCloud: (target: {
-    gitOwner: string;
-    gitRepoName: string;
-    ownerScope: AutomationRunOwnerScope;
-  }) => void;
-}) {
-  if (isLoading) {
-    return <RunLocationEmptyRow label="Loading targets" />;
-  }
-
-  const rows = groups.flatMap((group) => group.rows);
-  if (rows.length === 0) {
-    return <RunLocationEmptyRow label={emptyLabel} />;
-  }
-
-  return rows.map((row) => (
-    <RunLocationMenuRow
-      key={`${ownerScope}:${row.id}`}
-      activeOwnerScope={activeOwnerScope}
-      disabledReason={disabledReason}
-      ownerScope={ownerScope}
-      row={row}
-      onSelectOwner={onSelectOwner}
-      onSelectTarget={onSelectTarget}
-      onConfigureCloud={onConfigureCloud}
-    />
-  ));
-}
-
-function RunLocationMenuRow({
-  activeOwnerScope,
-  disabledReason,
-  ownerScope,
-  row,
-  onSelectOwner,
-  onSelectTarget,
-  onConfigureCloud,
-}: {
-  activeOwnerScope: AutomationRunOwnerScope;
-  disabledReason: string | null;
-  ownerScope: AutomationRunOwnerScope;
-  row: AutomationTargetRow;
-  onSelectOwner: (ownerScope: AutomationRunOwnerScope) => void;
-  onSelectTarget: (target: AutomationTargetSelection) => void;
-  onConfigureCloud: (target: {
-    gitOwner: string;
-    gitRepoName: string;
-    ownerScope: AutomationRunOwnerScope;
-  }) => void;
-}) {
-  if (row.kind === "configureCloud") {
-    return (
-      <RunLocationMenuItem
-        disabled={Boolean(disabledReason)}
-        icon={<Plus className="size-full" />}
-        label="Set up cloud"
-        detail={row.repoLabel}
-        title={disabledReason ?? row.description ?? undefined}
-        onClick={() => {
-          if (disabledReason) {
-            return;
-          }
-          onConfigureCloud({
-            gitOwner: row.gitOwner,
-            gitRepoName: row.gitRepoName,
-            ownerScope,
-          });
-        }}
-      />
-    );
-  }
-
-  const rowDisabledReason = disabledReason ?? row.disabledReason;
-  const selected = activeOwnerScope === ownerScope && row.selected;
-  return (
-    <RunLocationMenuItem
-      disabled={Boolean(rowDisabledReason)}
-      icon={targetRowIcon(row, "menu")}
-      label={row.label}
-      detail={row.repoLabel}
-      selected={selected}
-      title={rowDisabledReason ?? row.description ?? undefined}
-      onClick={() => {
-        if (rowDisabledReason) {
-          return;
-        }
-        onSelectOwner(ownerScope);
-        onSelectTarget(row.target);
-      }}
-    />
-  );
-}
-
-function RunLocationMenuItem({
-  detail,
-  disabled = false,
-  icon,
-  label,
-  onClick,
-  selected = false,
-  title,
-}: {
-  detail?: string | null;
-  disabled?: boolean;
-  icon?: ReactNode;
-  label: string;
-  onClick: () => void;
-  selected?: boolean;
-  title?: string;
-}) {
-  return (
-    <PopoverMenuItem
-      density="compact"
-      disabled={disabled}
-      title={title}
-      icon={icon}
-      label={(
-        <>
-          <span className="min-w-0 truncate">{label}</span>
-          {detail ? (
-            <span className="min-w-0 truncate text-muted-foreground">
-              {detail}
-            </span>
-          ) : null}
-        </>
-      )}
-      labelClassName="flex items-baseline gap-1.5 text-left"
-      trailing={selected ? <Check className="size-3.5" /> : null}
-      onClick={() => {
-        onClick();
-      }}
-    />
-  );
-}
-
-function RunLocationEmptyRow({ label }: { label: string }) {
-  return (
-    <div className="px-2 py-1 text-sm leading-4 text-muted-foreground">
-      {label}
-    </div>
-  );
-}
-
-function targetRowIcon(
-  row: Extract<AutomationTargetRow, { kind: "target" }>,
-  variant: "menu" | "trigger",
-) {
-  if (row.computeTargetOption) {
-    if (variant === "menu") {
-      return <ComputeTargetSwatch appearance={row.computeTargetOption.appearance} size="inherit" />;
-    }
-    return (
-      <span className="size-3.5">
-        <ComputeTargetSwatch appearance={row.computeTargetOption.appearance} size="inherit" />
-      </span>
-    );
-  }
-
-  const iconClassName = variant === "menu" ? "size-full" : "size-3.5";
-  if (row.target.executionTarget === "cloud") {
-    return <CloudIcon className={iconClassName} />;
-  }
-  if (row.target.executionTarget === "ssh") {
-    return <Terminal className={iconClassName} />;
-  }
-  return <FolderOpen className={iconClassName} />;
-}
-
-function findSelectedTargetRow(groups: AutomationTargetGroup[]) {
-  for (const group of groups) {
-    for (const row of group.rows) {
-      if (row.kind === "target" && row.selected) {
-        return row;
-      }
-    }
-  }
-  return null;
-}
-
-function findDefaultTargetRow(
-  groups: AutomationTargetGroup[],
-  preferredExecutionTarget?: AutomationTargetSelection["executionTarget"],
-) {
-  let fallback: Extract<AutomationTargetRow, { kind: "target" }> | null = null;
-  for (const group of groups) {
-    for (const row of group.rows) {
-      if (row.kind !== "target") {
-        continue;
-      }
-      if (!fallback && !row.disabledReason) {
-        fallback = row;
-      }
-      if (
-        preferredExecutionTarget
-        && row.target.executionTarget === preferredExecutionTarget
-        && !row.disabledReason
-      ) {
-        return row;
-      }
-    }
-  }
-  return fallback;
 }
