+++ apps/desktop/src/components/settings/panes/agent-defaults/AgentDefaultsSection.tsx
@@ -1,0 +1,23 @@
+import type { ReactNode } from "react";
+
+export function AgentDefaultsSection({
+  title,
+  description,
+  children,
+}: {
+  title: string;
+  description?: string;
+  children: ReactNode;
+}) {
+  return (
+    <div className="space-y-2">
+      <div className="space-y-0.5">
+        <h2 className="text-sm font-medium text-foreground">{title}</h2>
+        {description ? (
+          <p className="text-sm text-muted-foreground">{description}</p>
+        ) : null}
+      </div>
+      {children}
+    </div>
+  );
+}
