+++ scripts/max_lines_allowlist.txt
@@ -26,9 +26,6 @@ anyharness/sdk-react/src/hooks/sessions.ts 653 existing SDK React oversized file
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1880 existing SDK oversized test file
 anyharness/sdk/src/reducer/transcript.ts 1575 existing SDK oversized file
 apps/desktop/src-tauri/src/commands/ssh_tunnel.rs 729 existing desktop native oversized file
-apps/desktop/src/components/automations/editor/AutomationEditorModal.tsx 610 existing desktop component oversized file
-apps/desktop/src/components/settings/panes/SlackBotPane.tsx 545 existing desktop component oversized file
-apps/desktop/src/components/settings/panes/compute/ComputeTargetDetails.tsx 528 existing desktop component oversized file
 apps/desktop/src/components/workspace/chat/input/ComposerModelConfigSelector.tsx 735 existing desktop component oversized file
 apps/desktop/src/components/workspace/shell/right-panel/RightPanel.test.tsx 732 existing desktop component oversized test file
 apps/desktop/src/hooks/home/workflows/use-home-next-launch.ts 611 existing desktop hook oversized file
